Restore VSCode breadcrumbs sans symbols
It's nice to have the path nearby the tab, but the symbols after the
file path/name add a lot of noise that I never care about. The improved
sticky scrolling context makes some of that noise more redundant. This
restores the breadcrumbs in a cleaner format, then removes the file
path/name added to the window title to keep that clean/non-duplicative