Retain vscode/mas changes between installs
Previously, if an extension was uninstalled, and presumably if a mas app
was uninstalled, running the install script would result in
`brew bundle --global install` reinstalling the app. This updates the
setup script to check if the brew command already exists, and if so,
updates the Brewfile before installing from it. Otherwise, it will
freshly install everything from the Brewfile as expected (e.g. on a new
system)