Update dotfiles function to use 'commands'
In a similar way to git. This also absorbs the install_dotfiles function
into the main dotfiles function

The item in array checks are quite different to Bash. These helped:
- https://unix.stackexchange.com/a/411331/639529
- https://zsh.sourceforge.io/Doc/Release/Parameters.html#Subscript-Flags

read is also different, with the `-p` flag not being for prompts.
Instead you need to use a '?' as a separator in the argument:
- https://zsh.sourceforge.io/Doc/Release/Shell-Builtin-Commands.html#index-read
- https://stackoverflow.com/a/15174634/4002016

This is using some complex inline conditionals that I learned from bash
Idioms