Improve Starship prompt for vi bindings
This adds an indication of the mode you're in (insert vs
command/normal)

This means that the full prompt gets rewritten whenever you change mode,
and as a side effect the time module can't consistently represent when
the last command completed. As a workaround, we're using a custom module
to read and format the last entry of the history file. Sometimes the
date doesn't seem to be written to the history file (in which case it
prints a default 1970 date) but I haven't worked out the conditions (and
therefore can't consistently replicate it)..I might be able to make it
not show a date in this case, but that needs more investigation, and it
might be better to try setting EXTENDED_HISTORY instead
(https://zsh.sourceforge.io/Doc/Release/Options.html#History)

Note: I tried to get this to work with strftime or fc/history, but
neither quite worked correctly. Perhaps it's a non-interactive shell
issue (if the command runs in one)