Attempt to record history more consistently
I noticed that occasionally the timestamp of a command wasn't written to
the history file. I haven't worked out a way to replicate this/what the
conditions are, but there's a chance that this not being set explicitly
is the cause of inconsistent behaviour. I'm not sure why timestamps were
written without this (most of the time)