Fix rbenv/nodenv/Homebrew path position in tmux
tmux was using a login shell and therefore re-sourcing /etc/zprofile and
through that /etc/paths and /etc/paths.d/* when loaded. With that setup,
it would inherit the current path from the initial login shell, but then
re-run some path manipulation code in the second login shell, putting
the standard system directories back at the start of the path, ahead of
stuff like rbenv, nodenv, and Homebrew. The reason for not re-running
`rbenv init` and `nodenv init` in tmux was that they just prepend the
path again without removing duplicate entries

There are two ways this could be solved:
- continue using a login shell, revert to having rbenv, nodenv, and
  Homebrew prepend their directories to the path again (before the
  system directories), but add a step to remove duplicates after each
  path manipulation or at the end. This feels a little hacky, but
  perhaps the path of least resistance
- shift to using a non-login shell when launching tmux. This means
  setting the 'default command' that gets run at launch to just `zsh`,
  which is equivalent to running zsh from zsh, raising the shell level.
  This would prevent the system directories being prepended to the path
  after rbenv, nodenv, and Homebrew's directories have been added.
  .zshrc is still run from non-login shells, however, so we still need
  to avoid re-running `rbenv init` and `nodenv init` (and the Homebrew
  equivalent) to prevent duplicate entries

The second option is taken in this commit. A couple of new functions are
added to .zshenv for checking whether directories are in the path. This
allows us to prevent the duplication of entries both in tmux and any
other case where a zsh shell is launched from inside another zsh shell
(with a raised shell level)

To make the implications of the approach more visible, the shell level
is added to the prompt when it's two or higher

Note that the rbenv.zsh file was missing from .zshrc - that's fixed as
part of this commit