Change menu bar manager
Bartender 5 is broken on macOS Tahoe 26.0. There's a new major version
of Bartender but I've seen reports of it not behaving very well, so I'm
switching to Ice for now. The beta release contains various fixes for
Tahoe support - the main release didn't seem to work properly

I tried Ice before but decided to stick with Bartender back then. I
can't recall why, but it seems to be okay for now. It doesn't appear to
have the ability to group items, but I can live with that